test: cut a WebGL context and give slow beforeAll hooks headroom
CI failed with two `beforeAll` timeouts at 30s, in
`webgl_retained_geometry.spec.js` (new here) and
`texture-resource.spec.js` (pre-existing, untouched by this branch).
Both hooks do `boot()` + `video.init(WEBGL)`; 5282 tests passed, so
nothing was actually broken. Master is green, so the two spec files
added here pushed an already-marginal runner over.

Every spec file that calls `video.init` pays a WebGL context creation,
which on a container with no GPU goes through a software rasterizer.
The two specs added for #1507 needed one context between them, not two,
so `webgl_retained_geometry.spec.js` is folded into
`webgl_mesh_retained.spec.js` as a nested describe. Same 21 tests, one
context.

It has to sit before the context-loss test, not after: that test leaves
the context unusable, and running the unit tests behind it returned
CONTEXT_LOST_WEBGL from every `getError`. Same ordering trap the file
already documents.

`hookTimeout` also raised to 90s. The default fails a correct suite
whenever the container is loaded — this hook is genuinely slow, not
hanging, and `texture-resource.spec.js` timing out shows the margin was
already thin without this branch.
